Add launching simulation
Launching simulation is added as a simulation launch service. Simulation
is launched as a child process. The model can be run directly as a
standalone program or run as a script using the specified executable
(e.g., Python). Command line arguments passed to the model are generated
according to the predefined options in the command line argument parsing
services.

Additionally, since the dictionary storing all the predefined options in
the command line argument parsing services module is now also referenced
outside that module, it is changed such that it is no longer marked as
private.