Add parameter file check to parsing command line arguments
Checking if parameter file exists and is readable is added to the
command line argument parsing services.

Additionally, the name of the function that checks if a directory exists
and is writable is changed such that it is no longer marked as private
and also matches the name of the new function that checks if a file
exists and is readable.