Add simulation launcher
Simulation launcher is added as a console script. It launches model
simulation in a designated directory. It first creates an appropriate
directory structure including the simulation directory, generating
simulation id if necessary, then changes the current working directory
to the simulation directory, and finally launches a simulation as a
child process, passing all relevant command line arguments to the model.
Optionally, before launching the simulation, it can also copy the model
file as well as an optional parameter file to the simulation directory.