Add random seed generator
Random seed generator is added as a console script. It generates random
seed using OS-specific randomness source and then prints it. The
generated random seed can have a specified number of bytes or otherwise
its number of bytes is determined to match the number of bytes used by
the largest positive integer supported by the platform.