Add loading parameter names from a text file
Loading parameter names along with a mapping of parameter names from a
text file is added as a parameter service. The file must obey the
following format. Each line in the file can contain either a comment, in
which case it begins with the symbol "#", or a parameter name with an
optional substitute name. If an optional substitute name is provided, it
must come after the symbol "->" following the parameter name. Moreover,
empty lines in the file are ignored. If the file format is not obeyed,
an exception is raised.

Additionally, since the exception that can be raised during processing a
file with names of parameters is very similar to the one that can be
raised during processing a parameter file, the latter exception is
adapted to serve both purposes.