Add loading names of simulation directories from a text file
Loading names of simulation directories from a text file is added as a
simulation launch service. The file must obey the following format. Each
line in the file can contain a comment, in which case it begins with the
symbol "#", or a single directory name, or a directory path. In the case
of a directory path, both slashes and backslashes can be used. Moreover,
empty lines in the file are ignored.