Add parameter exporter
Parameter exporter is added as a console script. It exports parameters
of multiple simulations to a file. It first loads from a text file the
names of parameters to be exported along with an optional mapping that
defines how these names should be substituted with different ones, then
loads from another text file the names of relevant simulation
directories, and finally, as the parameters are collected by traversing
the simulation directories and loading appropriate parameter files, it
exports them to a file.

Additionally, since specifying a simulation master directory as an
optional argument necessitates validation that it exists and is
readable, a function that checks if a directory exists and is readable
is added to the command line argument parsing services.