Add handling extra arguments to parsing command line arguments
Handling optional extra arguments is added to parsing command line
arguments. To match the argument parser in the Python Standard Library,
two cases of parsing command line arguments are distinguished and each
of them is assigned to a different function: one that does not allow
extra arguments and the other one that allows extra arguments. The
common code is moved to a separate private function. The function that
allows extra arguments passes them separately from the populated
supported options.