Add using specified parser to parsing command line arguments
Using an optionally specified parser for parsing command line arguments
is added to parsing command line arguments. Such an externally provided
parser can be utilized to define additional arguments to be parsed along
with arguments used to populate supported options. The descriptions of
the arguments corresponding to supported options are thus supplied to
the parser on top of those already defined there, which implies that the
names of the additional arguments must not conflict with those of the
arguments corresponding to the supported options.