Add using only long names of arguments to parsing command line arguments
Using optionally only the long names of the arguments corresponding to
supported options is added to parsing command line arguments. Applying
this restriction may reduce the likelihood of conflicts between the
names of additional arguments utilized in the model file (for example,
through an externally provided parser) and the names of the arguments
corresponding to supported options.

Additionally, to ensure full support for this feature, launching
simulation provided in the simulation launch services module is changed
to use only the long names of the arguments corresponding to supported
options when generating a command to be executed for launching
simulation.